protocol: add MockClient
