+++ packages/indexer/src/vcr/index.ts
@@ -1,4 +1,3 @@
-export * from "./client";
 export * from "./config";
 export * from "./helper";
 export * from "./record";
