+++ packages/protocol/package.json
@@ -4,7 +4,10 @@
   "type": "module",
   "source": "./src/index.ts",
   "main": "./src/index.ts",
-  "exports": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./testing": "./src/testing/index.ts"
+  },
   "publishConfig": {
     "files": ["dist", "src", "README.md"],
     "main": "./dist/index.mjs",
@@ -14,6 +17,11 @@
         "types": "./dist/index.d.ts",
         "import": "./dist/index.mjs",
         "default": "./dist/index.mjs"
+      },
+      "./testing": {
+        "types": "./dist/testing/index.d.ts",
+        "import": "./dist/testing/index.mjs",
+        "default": "./dist/testing/index.mjs"
       }
     }
   },
